Added metadata to all pages for search engine optimisation - Projects page decomposed to make it a server component
