+++ app/blogs/page.tsx
@@ -1,7 +1,13 @@
 import getBlogMetadata from "@/actions/getBlogMetadata";
 import HeadingOne from "@/components/Text/HeadingOne";
 import { BlogList } from "./components/BlogList";
+import type { Metadata } from "next";
 
+export const metadata: Metadata = {
+  title: "Maruf - Blogs",
+  description:
+    "This is the blog page for my personal website. It contains all the blogs I have written. ",
+};
 /**
  * Displays a list of all blogs that can be opened.
  * Also allows the user to search for blogs.
